+++ spec/ruby/core/process/clock_gettime_spec.rb
@@ -1,6 +1,12 @@
 require_relative '../../spec_helper'
 
 describe "Process.clock_gettime" do
+  it 'can be called with all declared clocks' do
+    Process.constants.select { |c| c.to_s.start_with?('CLOCK_') }.each do |c|
+      Process.clock_gettime(Process.const_get(c)).should be_an_instance_of(Float)
+    end
+  end
+
   describe 'time units' do
     it 'handles a fixed set of time units' do
       [:nanosecond, :microsecond, :millisecond, :second].each do |unit|